Settings Menu Implemented
Still need to get variables to persist more steadily between restarts.
And the code is insanely messy in parts so I need to clean that up.

Chat log dump is possible but I'll have to add another aspect of the
addon to constantly read in the chat logs, rather than using an API to
dump it out for me.