cleanliness + documentation + features = damn
Actually added in persistent variables this time. Oh and the code isn't
complete shit to look at now. It's pretty well organized (except for the
UI elements, but don't scroll down that far) and well commented now. I
wouldn't be that embarassed to show someone it in this stage, even
though it still is pretty rudimentary

As for actual features/changes:
The 'output' file we are using contains ANY persistent variables
contained within a table. So now you will notice a new variables = {}
table in addition to the ssdatabase = {} table. it will probably fuck
your interpretor up so let me know how I can make it easier for you to
skip over.

I also added in logs for ends of sessions for characters so we can
calculate sessions if we really wanted to in the future.
"SESSION START AT 02/28/17 17:48:22 SAVED TO
../SavedVariables/SStagger.lua.. ", -- [127]
LOGS THAT HE WOULDFVE MADE HERE
"PLAYER Moslenick SESSION ENDED AT 02/28/17 17:48:37", -- [129]

This also will probably fuck up your interpretor so let me know if you
want me to add delimiters or what to help you out.