Operation UNFUCK IT UP
What happened is the user config file was never found on your computer
so it didnt load correctly. It should be fixed now in the case where
there is no UI parameters already stored on your comp, so it will
properly build the default file.

It worked for me because I already had a config file table in the .lua
log output file, but it would NOT work if this config file table was not
in your output file.

LEMME KNOW IF IT WORKS